app/requirements.txt
Django==5.2.5
gunicorn
psycopg2
djangorestframework
pytest
pytest-django